Added rudimentary usb handling functions and made linking safer
There is now a restriction preventing .text from accessing .bootloader. This
prevents any unintentional function calls. There is now a .bootstrap section
which acts as the bridge between .bootloader and .text.

We are also having memcpy issues...the additional union in usb_types.h seems
to have stopped that for now, but it could come back.